pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oauthkitchen"
version = "0.1.0"
description = "OAuth Security Analysis Tool for Microsoft Entra ID - Consent Analyzer, Shadow OAuth Detector, Permission Translator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "OAuthKitchen Contributors" }
]
keywords = [
    "oauth",
    "microsoft",
    "entra",
    "azure-ad",
    "security",
    "graph-api",
    "consent",
    "permissions"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "msal>=1.24.0",
    "requests>=2.31.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "types-python-dateutil>=2.8.0",
]

[project.scripts]
oauthkitchen = "oauthkitchen.cli:main"

[project.urls]
Homepage = "https://github.com/oauthkitchen/oauthkitchen"
Documentation = "https://github.com/oauthkitchen/oauthkitchen#readme"
Repository = "https://github.com/oauthkitchen/oauthkitchen"
Issues = "https://github.com/oauthkitchen/oauthkitchen/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/oauthkitchen"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert usage (needed for tests)
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=oauthkitchen --cov-report=term-missing"
